DEV: Remove obsolete parts of our custom loader
