+++ .prettierignore
@@ -9,7 +9,6 @@ config/locales/**/*.yml
 script/import_scripts/**/*.yml
 
 app/assets/javascripts/browser-update.js
-app/assets/javascripts/discourse-loader.js
 app/assets/javascripts/locales/i18n.js
 app/assets/javascripts/ember-addons/
 app/assets/javascripts/discourse/lib/autosize.js
